+++ src/main/java/uk/co/jpawlak/thelongdark/mapnotes/serializable/SettingsSerialiser.java
@@ -6,12 +6,18 @@ import java.io.File;
 
 public class SettingsSerialiser {
 
+    private static final File SETTINGS_FILE = new File(Main.MAIN_FOLDER, "settings.json");
+
     public static void save(Settings settings) {
         Serialiser.save(settings, Main.MAIN_FOLDER, "settings");
     }
 
+    public static boolean settingsExist() {
+        return SETTINGS_FILE.exists();
+    }
+
     public static Settings load() {
-        return Serialiser.load(new File(Main.MAIN_FOLDER, "settings.json"), Settings.class);
+        return Serialiser.load(SETTINGS_FILE, Settings.class);
     }
 
 }